+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 ### Bug Fixes
 
+* `equals()` and `hashCode()` of managed `RealmObject`s that come from linking objects don't work correctly (#4487).
+
 ### Internal
 
 * Upgraded to Realm Sync 1.5.2.